pull main branch

+++ src/main/java/edu/lucasrech/catalog_authentication/service/ProductService.java
@@ -1,9 +1,8 @@
 package edu.lucasrech.catalog_authentication.service;
 
 import edu.lucasrech.catalog_authentication.model.Product;
-import edu.lucasrech.catalog_authentication.model.enums.Category;
 import edu.lucasrech.catalog_authentication.repository.ProductRepository;
-import org.springframework.beans.factory.annotation.Autowired;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -11,8 +10,11 @@ import java.util.List;
 @Service
 public class ProductService {
 
-    @Autowired
-    private ProductRepository productRepository;
+    private final ProductRepository productRepository;
+
+    public ProductService(ProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
 
     public List<Product> getAllProducts() {
         return productRepository.findAll();
@@ -27,14 +29,26 @@ public class ProductService {
     }
 
     public Product updateProduct(String id, Product product) {
-        return productRepository.findById(id).orElse(null);
+        Product existingProduct = productRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Product not found"));
+        existingProduct.setName(product.getName());
+        existingProduct.setDescription(product.getDescription());
+        existingProduct.setPrice(product.getPrice());
+        existingProduct.setCategory(product.getCategory());
+        existingProduct.setImageUrl(product.getImageUrl());
+
+        return productRepository.save(existingProduct);
     }
 
     public void deleteProduct(String id) {
+        if (!productRepository.existsById(id)) {
+            throw new EntityNotFoundException("Product with id " + id + " not found");
+        }
         productRepository.deleteById(id);
+
     }
 
-    public List<Product> getProductsByCategory(Category category) {
+    public List<Product> getProductsByCategory(String category) {
         return productRepository.findByCategory(category);
     }
